Cache results of the patternWithString. Clear entire cache when it reaches 150 entries (currently Robinhood app is using 30 entries).
Having this cache further reduces memory footprint spikes
and should help with performance a bit (not measured).

Use OSSpinlock to synchronize between threads.
OSSPinlock achieve best performance on the same CPU thread.
OSSpinlock are 10x faster than synchronize, but have a drawback
of not yielding. For this piece of code, yielding itself would be
a drawback, thus OSSpinlock is the best synchronization tool we could use

More info:
http://perpendiculo.us/2009/09/synchronized-nslock-pthread-osspinlock-showdown-done-right/